move middleware to calender
Fixes redirect issues, no need for exception routes
