--- routes/_middleware.ts
@@ -1,74 +1,0 @@
-import { MiddlewareHandlerContext } from "$fresh/server.ts";
-import {
-  AUDIO_LOGBOOK_AUTH_COOKIE_NAME,
-  DEPLOYMENT_ID,
-} from "@/src/const/server_constants.ts";
-import { UserSession } from "@/src/db/db_schema.ts";
-import { secretsPromise } from "@/src/utils/secrets.ts";
-import { getCookies } from "$std/http/cookie.ts";
-import { dbPromise } from "@/src/db/db.ts";
-import { gotoLogin } from "@/src/utils/redirects.ts";
-import { ContextState } from "@/src/context_state.ts";
-
-export async function handler(
-  req: Request,
-  ctx: MiddlewareHandlerContext<ContextState>,
-) {
-  // Initialize serverOrigin deferred promise
-  const origin = new URL(req.url).origin;
-  ctx.state.serverOrigin = origin;
-
-  // ----------------------------------------
-  // Handle special requests
-
-  // Do not alter requests to the telegram bot
-  if (new URL(req.url).pathname.startsWith("/bot")) {
-    // console.debug(`Bot Request, doing nothing in middleware`);
-    return ctx.next();
-  }
-
-  // If user is on the login page, do not redirect to login (bc.this would be endless loop)
-  if (new URL(req.url).pathname.startsWith(`/auth/login`)) {
-    return ctx.next();
-  }
-
-  // Special requests end here
-  // ----------------------------------------
-
-  // When in dev mode,
-  // - get auth token from doppler secrets for local mocking, or
-  // - get auth token from request header in anything other than dev mode
-  // - and parse it
-  const url = new URL(req.url);
-  const secrets = await secretsPromise;
-  const maybeAccessToken =
-    (secrets.get("ENV_NAME") === "dev" || url.host.includes(DEPLOYMENT_ID))
-      ? secrets.get("MOCK_AUTH_TOKEN")
-      : getCookies(req.headers)[AUDIO_LOGBOOK_AUTH_COOKIE_NAME]; //return undefined if no cookie is set
-
-  if (maybeAccessToken) {
-    const db = await dbPromise;
-    const maybeUserQuery = db
-      .selectFrom("audiobook_sessions")
-      .selectAll()
-      .where(
-        "hash",
-        "=",
-        maybeAccessToken,
-      );
-    const maybeUser = await maybeUserQuery.executeTakeFirst();
-    const userParse = UserSession.safeParse(maybeUser);
-
-    // Redirect to login page when UserSession is not available
-    if (!userParse.success) {
-      return gotoLogin();
-    }
-
-    // Happy Path to the normal App
-    ctx.state.user = userParse.data;
-    return ctx.next();
-  }
-
-  // Redirect to login page when Auth Cookie is not available
-  return gotoLogin();
-}
